+++ app/Http/Controllers/OrganizationController.php
@@ -6,39 +6,48 @@ use App\Http\Resources\OrganizationResource;
 use App\Models\Organization;
 use Exception;
 use Illuminate\Http\Request;
-use PhpParser\Node\Expr;
 
 class OrganizationController extends Controller
 {
     public function index(){
-        $organization = Organization::with('childrens')->where('parent_id',0)->first();
-        // dd($organization);
-        return view('organizations.departments.index',compact('organization'));
+        $this->authorize('viewany',Organization::class);
+        return view('organizations.departments.index');
     }
 
     public function getOrganizationsAjax(){
-        $organizations = OrganizationResource::collection(Organization::all());
+
+        if(auth()->user()->type === 'super_admin'){
+            $organizations = Organization::all();
+        }else{
+            $organizations = Organization::where('id',auth()->user()->organization_id)->get();
+        }
+        $organizations = OrganizationResource::collection($organizations);
         return response()->json(['organizations'=>$organizations]);
     }
 
     public function store(Request $request)
     {
         // return $request->all();
+        if(!auth()->user()->can('create',Organization::class)){
+            return response()->json(['message','UnAuthorized '],403);
+           }
         $request->validate([
             'name' =>"required",
-            'description' => 'required',
+            // 'description' => 'required',
             'parent_id'   => 'required'
         ]);
 
         //store organization;
         try{
-
-            $organization= Organization::create($request->all());
+            $organization_data = array(
+                'name' => $request->input('name'),
+                'description' => $request->input('description') ?? " ",
+                'parent_id' => $request->input('parent_id'),
+            );
+            $organization= Organization::create($organization_data);
             return response()->json(['organization'=>$organization]);
-            // $this->success('Organization has been created successfully');
         }catch(Exception $e){
-            $this->error($e->getMessage());
-           return redirect()->back();
+            return response()->json(['error'=>$e->getMessage()],500);
         }
     }
 
@@ -49,20 +58,23 @@ class OrganizationController extends Controller
             $this->error("Organization Not Found");
             return redirect()->back();
         }
+        if(!auth()->user()->can('update',$organization)){
+            return response()->json(['message','UnAuthorized '],403);
+        }
         $request->validate([
-            'name' =>"required",
-            'description' => 'required',
-            // 'parent_id'   => 'required'
+            'name'        =>  "required",
         ]);
 
          //store organization;
          try{
-            $organization->update($request->all());
+            $organization_data = array(
+                'name' => $request->input('name'),
+                'description' => $request->input('description') ?? " ",
+            );
+            $organization->update($organization_data);
             return response()->json(['organization'=>$organization]);
-            // $this->success('Organization has been created successfully');
         }catch(Exception $e){
-            // $this->error($e->getMessage());
-           return response()->json($e->getMessage());
+            return response()->json(['error'=>$e->getMessage()],500);
         }
 
     }
@@ -73,13 +85,15 @@ class OrganizationController extends Controller
         if(!$organization){
             return response()->json(['error'=>"Organization Not Found"],404);
         }
+        if(!auth()->user()->can('delete',$organization)){
+            return response()->json(['message','UnAuthorized '],403);
+        }
         try{
             if($organization->parent_id === 0){
                 return response()->json(['error'=>"Can't Delete Parent Organization"],404);
             }
           $organization->delete();
           return response()->json(['success'=>"Organization Deleted Successfully"]);
-        //   $this->success();
         }catch(Exception $e){
             $this->error($e->getMessage());
         }
